#22 refactor: 관심사 분리 및 input 구현
