+++ src/shared/ui/Input/Email/index.tsx
@@ -1,6 +1,6 @@
-import Label from "./Label";
-import Input from ".";
-import HelperText from "./HelperText";
+import Label from "../Label";
+import Input from "..";
+import HelperText from "../HelperText";
 import { AuthInputProps } from "@/shared/types/input";
 
 const EmailInput = ({ register, errors }: AuthInputProps) => {
@@ -8,19 +8,17 @@ const EmailInput = ({ register, errors }: AuthInputProps) => {
     <div className="flex flex-col gap-2.5">
       <Label>이메일</Label>
       <Input
-        id="email"
         inputSize="large"
         type="email"
         placeholder="이메일을 입력해주세요"
-        register={register}
-        errors={errors}
-        validation={{
+        {...register("email", {
           required: "이메일을 입력해주세요",
           pattern: {
             value: /^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$/,
             message: "잘못된 이메일입니다",
           },
-        }}
+        })}
+        isError={!!errors.email}
       />
       {errors.email && (
         <HelperText type="error">{errors.email.message}</HelperText>
