+++ src/shared/ui/Input/Nickname/index.tsx
@@ -1,6 +1,6 @@
-import Label from "./Label";
-import Input from ".";
-import HelperText from "./HelperText";
+import Label from "../Label";
+import Input from "..";
+import HelperText from "../HelperText";
 import { AuthInputProps } from "@/shared/types/input";
 
 const NicknameInput = ({ register, errors }: AuthInputProps) => {
@@ -8,19 +8,17 @@ const NicknameInput = ({ register, errors }: AuthInputProps) => {
     <div className="flex flex-col gap-2.5">
       <Label>닉네임</Label>
       <Input
-        id="nickname"
         inputSize="large"
         type="text"
         placeholder="닉네임을 입력해 주세요"
-        register={register}
-        errors={errors}
-        validation={{
+        {...register("nickname", {
           required: "닉네임을 입력해주세요",
           maxLength: {
             value: 10,
             message: "닉네임은 10자 이하로 작성해주세요",
           },
-        }}
+        })}
+        isError={!!errors.nickname}
       />
       <HelperText type={errors.nickname ? "error" : "basic"}>
         {errors.nickname ? errors.nickname.message : "최대 10자 가능"}
