+++ src/shared/ui/Input/Password/index.tsx
@@ -1,9 +1,9 @@
 import { useState } from "react";
-import Label from "./Label";
-import Input from ".";
-import HelperText from "./HelperText";
+import Label from "../Label";
+import Input from "..";
+import HelperText from "../HelperText";
 import { AuthInputProps } from "@/shared/types/input";
-import Icon from "../Icon/Icon";
+import Icon from "../../Icon/Icon";
 
 const PasswordInput = ({ register, errors }: AuthInputProps) => {
   const [isShowPassword, setIsShowPassword] = useState(false);
@@ -17,13 +17,10 @@ const PasswordInput = ({ register, errors }: AuthInputProps) => {
       <Label>비밀번호</Label>
       <div className="relative">
         <Input
-          id="password"
           inputSize="large"
           type={isShowPassword ? "text" : "password"}
           placeholder="비밀번호를 입력해주세요"
-          register={register}
-          errors={errors}
-          validation={{
+          {...register("password", {
             required: "비밀번호를 입력해주세요",
             minLength: {
               value: 8,
@@ -34,7 +31,8 @@ const PasswordInput = ({ register, errors }: AuthInputProps) => {
                 /^(?=.*[A-Za-z])(?=.*\d)(?=.*[!@#$%^&*()_\-+=<>?{}[\]~]).{8,}$/,
               message: "비밀번호는 영어, 숫자, 특수기호를 포함해야 합니다.",
             },
-          }}
+          })}
+          isError={!!errors.password}
         />
         <Icon
           name={isShowPassword ? "VisibilityIcon" : "VisibilityOffIcon"}
@@ -42,9 +40,9 @@ const PasswordInput = ({ register, errors }: AuthInputProps) => {
           onClick={toggleIcon}
         />
       </div>
-      {errors.password && (
-        <HelperText type="error">{errors.password.message}</HelperText>
-      )}
+      <HelperText type={errors.password ? "error" : "basic"}>
+        {errors.password ? errors.password.message : "최소 8자 이상"}
+      </HelperText>
     </div>
   );
 };
